Add old python toolkit package
This is the old version of "python-toolkit" that I had made while working on
Project Euler during 2017. Very few of these functions had any proper profiling
done, any they are all subject to change or removal from the package. As of
this commit, the package is not in a satisfactory state.